% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_flob}
\alias{check_flob}
\title{Check flob}
\usage{
check_flob(x, old = FALSE, x_name = substitute(x))
}
\arguments{
\item{x}{The object to check.}

\item{old}{A flag indicating whether old flobs are permitted.}

\item{x_name}{A string of the object name.}
}
\value{
An invisible copy of x.
}
\description{
Checks whether an object is a \code{\link{flob}}.
}
\examples{
check_flob(flob_obj)
}
\seealso{
\code{\link{flobr}}
}
