% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flob_ext}
\alias{flob_ext}
\title{Flob Extension}
\usage{
flob_ext(flob)
}
\arguments{
\item{flob}{A \code{\link[=flob]{flob()}}}
}
\value{
A string of the extension of the original file.
}
\description{
Flob Extension
}
\examples{
flob_ext(flob_obj)
}
\seealso{
\code{\link[=flobr]{flobr()}} and \code{\link[=flob_name]{flob_name()}}.
}
