% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flob_name}
\alias{flob_name}
\title{Flob Name}
\usage{
flob_name(flob)
}
\arguments{
\item{flob}{A \code{\link{flob}}}
}
\value{
A character vector of the name of the file (without the extension).
}
\description{
Gets a character vector of the name of the file (without the extension).
If the flob does not have a file name it returns "file".
}
\examples{
flob_name(flob_obj)
}
\seealso{
\code{\link{flobr}} and \code{\link{flob_ext}}.
}
