% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flob_obj}
\alias{flob_obj}
\title{A flob Object}
\format{An object of class \code{flob} (inherits from \code{blob}) of length 1.}
\usage{
flob_obj
}
\description{
A \code{\link{flob}} of \code{'inst/extdata/flobr.pdf'}.
}
\examples{
flob_obj
}
\keyword{datasets}
