% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_flob}
\alias{is_flob}
\title{Is flob}
\usage{
is_flob(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A flag indicating whether x inherits from \code{\link[=flob]{flob()}}.
}
\description{
Is flob
}
\examples{
is_flob(flob_obj)
}
\seealso{
\code{\link[=flobr]{flobr()}}
}
