% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{pkgtemplate_deprecated}
\alias{pkgtemplate_deprecated}
\alias{check_flob}
\title{Deprecated Functions}
\usage{
check_flob(x, old = FALSE, x_name = NULL)
}
\description{
Deprecated \code{pkgtemplate} functions.
}
\section{Functions}{
\itemize{
\item \code{check_flob}: Check Function Template

\lifecycle{soft-deprecated}
}}

\examples{

check_flob(flobr::flob_obj)
}
\keyword{internal}
