% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vld.R
\name{vld_blob}
\alias{vld_blob}
\title{Validate blob Object}
\usage{
vld_blob(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A flag indicating whether the object passed the test.
}
\description{
Validates a \code{\link[=blob]{blob()}} object.
}
\seealso{
\code{\link[=chk_blob]{chk_blob()}}
}
