% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vld.R
\name{vld_flob}
\alias{vld_flob}
\title{Validate flob Object}
\usage{
vld_flob(x, old = FALSE)
}
\arguments{
\item{x}{The object to check.}

\item{old}{A flag indicating whether old flobs are permitted.}
}
\value{
A flag indicating whether the object passed the test.
}
\description{
Validates a \code{\link[=flob]{flob()}} object.
}
\examples{
vld_flob(flobr::flob_obj)
vld_flob(1)
}
\seealso{
\code{\link[=chk_flob]{chk_flob()}}
}
