% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flob.R
\name{flob}
\alias{flob}
\title{Flob File}
\usage{
flob(path)
}
\arguments{
\item{path}{A string of the path to the file.}
}
\value{
A named flob of the file.
}
\description{
Converts a file into a flob.
}
\details{
A flob is a serialized blob of a file,
which was read into binary in integer-mode as little endian,
that also includes the extension of the original file.
}
\examples{
path <- system.file("extdata", "flobr.pdf", package = "flobr")
flob <- flob(path)
flob
is_flob(flob)
flob_ext(flob)
unflob(flob, tempfile())
}
\seealso{
\code{\link{blob}}, \code{\link{serialize}}, \code{\link{readBin}}
and \code{\link{unflob}}
}
