% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flobr-package.R
\docType{package}
\name{flobr-package}
\alias{flobr}
\alias{flobr-package}
\title{flobr: Convert a file to and from binary data (BLOB)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca}

}
\keyword{internal}
