% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flobr-package.R
\docType{package}
\name{flobr-package}
\alias{flobr}
\alias{flobr-package}
\title{flobr: Convert a file to and from binary data (BLOB)}
\description{
Converts a file to and from a flob. 
    A flob is a serialized blob of a file,
    which was read into binary in integer-mode as little endian, 
    that also includes the extension of the original file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/flobr}
  \item Report bugs at \url{https://github.com/poissonconsulting/flobr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca}

}
\keyword{internal}
