% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flobr-package.R
\docType{package}
\name{flobr-package}
\alias{flobr}
\alias{flobr-package}
\title{flobr: Convert a file to and from binary data (BLOB)}
\description{
Converts a file to and from a flob. 
    A flob is a file that was 
    read into binary in integer-mode as little endian, 
    saved as the single element of a named list where the name is the extension
    of the original file and then serialized before being coerced into a blob.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/flobr}
  \item Report bugs at \url{https://github.com/poissonconsulting/flobr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)

}
\keyword{internal}
