% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flob.R
\name{unflob}
\alias{unflob}
\title{Unflob to File}
\usage{
unflob(flob, path)
}
\arguments{
\item{flob}{The \code{\link{flob}} to save to file.}

\item{path}{A string of the path to the new file.}
}
\value{
An invisible string of the path to the saved file.
}
\description{
Converts a \code{\link{flob}} back to its original file format.
}
\details{
If path does not include an extension then it is added automatically.
If path does include an extension and then it must match the flob's extension.
}
\examples{
unflob(flob_obj, tempfile())
}
